move healpix_ra_dec to utils.
