+++ api/fastpm/utils.h
@@ -25,5 +25,12 @@ fastpm_utils_load(PM * pm , const char * filename, FastPMFloat *data);
 double 
 fastpm_utils_get_random(uint64_t id);
 
+void
+fastpm_utils_healpix_ra_dec (
+                size_t nside,
+                double **ra,
+                double **dec,
+                size_t * npix);
+
 FASTPM_END_DECLS
