+++ libfastpm/chealpix.c
@@ -34,7 +34,7 @@
 // #ifdef ENABLE_FITSIO
 // #include "fitsio.h"
 // #endif
-#include <fastpm/chealpix.h>
+#include "chealpix.h"
 
 static const double twothird=2.0/3.0;
 static const double pi=3.141592653589793238462643383279502884197;
