+++ libfastpm/utils.c
@@ -7,6 +7,7 @@
 #include <fastpm/logging.h>
 #include <fastpm/string.h>
 #include <fastpm/transfer.h>
+#include "chealpix.h"
 
 #include "pmpfft.h"
 #include "pmghosts.h"
@@ -207,5 +208,29 @@ double
 fastpm_utils_powerspec_white(double k, double * amplitude) 	/* White Noise*/
 {
     return *amplitude;
+}
+
+void
+fastpm_utils_healpix_ra_dec (
+                size_t nside,
+                double **ra,
+                double **dec,
+                size_t * npix)
+{
+    const double rad_to_degree = 180./M_PI;
+
+    *npix = nside2npix (nside);
+    //fastpm_info("healpix npix %ld \n",*npix);
+    *ra = malloc(sizeof(double) * *npix);
+    *dec = malloc(sizeof(double) * *npix);
+    size_t i=0;
+    for (i = 0;i < *npix; i++)
+    {
+        pix2ang_ring(nside,i,&((*dec)[i]),&((*ra)[i]));
+        (*ra)[i]*=rad_to_degree;
+        (*dec)[i]*=rad_to_degree;
+        (*dec)[i]= 90 - (*dec)[i];
+        //fastpm_info("healpix radec %ld %g %g \n",i,ra[0][i],dec[0][i]);
+    }
 }
